requirements.txt
appdirs==1.4.3
asgiref==3.2.7
certifi==2020.4.5.1
distlib==0.3.0
Django==3.0.6
django-mdeditor==0.1.18
filelock==3.0.12
gunicorn==20.0.4
Markdown==3.2.1
mysqlclient==1.4.6
Pillow==7.1.2
pipenv==2018.11.26
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
virtualenv==20.0.20
virtualenv-clone==0.5.4
xlwt==1.3.0
